requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
blinker==1.6.2
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.7
cryptography==38.0.3
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.1.0
faiss-cpu==1.7.4
filelock==3.12.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.16.0
ipython==8.5.0
ipywidgets==8.0.6
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-core==4.11.1
jupyter-ui-poll==0.2.2
jupyter_client==7.3.5
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langchain==0.0.194
langchainplus-sdk==0.0.6
llvmlite==0.40.1rc1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
msal==1.20.0
msal-extensions==1.0.0
msedge-selenium-tools==3.141.4
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.0
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.6
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
portalocker==2.7.0
prompt-toolkit==3.0.31
protobuf==4.23.3
psutil==5.9.2
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
pygame==2.1.2
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.4.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==3.141.0
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.5.1
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tmdbsimple==2.9.1
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
typing-inspect==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
uamqp==1.6.1
urllib3==1.26.12
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
